pyproject.toml
[tool.isort]
profile = "black"

[tool.poetry]
name = "scicamera"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.23.5"
pillow = "^9.3.0"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
isort = "^5.11.2"
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
